- Fix inclusion of primary key in edit forms
When the PK of the model is a serial field (ID), it’s necessary to
include it in the declaration of the XML View as a hidden column
(display = "none") in order to access the data when recording model.
The column is now included as hidden, thus being included in the
recording process.