Add subtitle to extended controllers
Added function "primaryDescription" in the models returned by the
Descriptive identifier for humans of the data record.

Added function "getPrimaryDescription" to the data editing controllers EditController and PanelController, which returns the descriptive identifier of the main record that is being edited.

Modified the EditController and PanelController templates so that they visualize the descriptive identifier of the record that is being edited.