Fix WidgetSelect auto load options
The automatic loading of values from a table in the database failed because a item was added to the array of values before processing the "auto values". The component expects that in the first element of the array of values are the data necessary to obtain the "auto values". The load of this new item is moved to after the auto charge.