Minor fix: Remove HTML Tags before resume text
Is required to remove HTML tags before return the content af the string, because if really contains HTML tags, results on incomplete code extract that break ower code.

In both cases, return a version without HTML tags, because we don't need to see as HTML that can modify base template if it is not removed. When only want to see if has content setted.