Fix Options css style and translate title in WidgetSelect
The Select widget now applies the css styles indicated with the "option" tags.

Now translation is applied to the titles when the values are loaded from an array and the "title" is informed. If "title" is not reported, the "value" content is displayed as a description without applying translation.

For cases in which titles and values are loaded from the database, translation is applied if the attribute "translate" is added to the declaration of the "value" tag with value "true".