Add numcolumns, class and body to render method
The label attribute is changed by body and its rendering is added.
The numcolumns attribute is added in order to indicate the size to be used by the card.
The class attribute is added to be able to indicate css classes.
The color attribute is discarded because it can be done using the class attribute.