Added Bank charges
The special account for bank charges is added.
The bank expense account is added to the bank accounts of the company.
The sub-account widget is modified in the edition of bank accounts so that it is autocomplete and thus can be searched by account and by description.
The sub-account and sub-account columns for bank charges are added to the list view of bank accounts.