Agent system (New Version)
Added fiscal identifier to the tables and models of Agent and Contact.
Refactoring the agent system to use the contact table instead of containing the address data in the agent model itself.
Modified the EditContacto view to show the tipoidfiscal and codagente fields.
Modified the EditAgenteController so that it extends from the new EditDocHistoryController class.
Added contact view to the EditAgente controller.