Fix Customer's risk updated
Until now, if a user updated a customer's invoice receipt, the customer's risk remained the same. With this update the risk is updated. The risk calculation formula has also been changed in order to have a more realistic approximation of the risk. Now the formula is:
Risk = Cost of new orders + Cost of undelivered delivery notes + ( Amount of unpaid invoices - Amount of paid receipts of unpaid invoices).